+++ pyhapi/hapi.py
@@ -552,7 +552,7 @@ async def wait_cook_async(session, status_report_interval=0.1, status_verbosity=
             Defaults to 1.
     """
     if status_verbosity>HDATA.StatusVerbosity.WARNINGS:
-        logging.info("-------------Start Cooking!---------------")
+        logging.info("-------------Start  Cooking at Session {0}!---------------".format(session.id))
     cook_status = c_int32()
     cook_result = HDATA.Result.ALREADY_INITIALIZED
     while True:
@@ -561,7 +561,7 @@ async def wait_cook_async(session, status_report_interval=0.1, status_verbosity=
         continuestate = cook_status.value > HDATA.State.MAX_READY_STATE\
             and cook_result == HDATA.Result.SUCCESS
         if status_verbosity>HDATA.StatusVerbosity.WARNINGS:
-            logging.info("Cook Status at {0} : {1}".format(datetime.now().\
+            logging.info("Cook Status at session {0} {1}: {2}".format(session.id, datetime.now().\
                 strftime('%H:%M:%S'), _get_status_string(session,\
                 HDATA.StatusType.COOK_STATE,\
                     status_verbosity)))
@@ -569,13 +569,13 @@ async def wait_cook_async(session, status_report_interval=0.1, status_verbosity=
             break
         await asyncio.sleep(status_report_interval)
     if cook_status.value == HDATA.State.READY_WITH_FATAL_ERRORS:
-        logging.error("Cook with Fatal Error: {0}".format(_get_status_string(session)))
+        logging.error("Cook with Fatal Error at session {0} {1}".format(session.id, _get_status_string(session)))
     if status_verbosity>HDATA.StatusVerbosity.WARNINGS:
-        logging.info("-------------Finish Cooking!---------------")
+        logging.info("-------------Finish Cooking at Session {0}!---------------".format(session.id))
     assert cook_result == HDATA.Result.SUCCESS and\
         cook_status.value == HDATA.State.READY,\
-        "CookNode Failed with {0} and Cook Status is {1}".\
-        format(HDATA.Result(cook_result).name,
+        "CookNode Failed with {0} at session {1} and Cook Status is {2}".\
+        format(HDATA.Result(cook_result).name, session.id, 
                HDATA.State(cook_status.value).name)